Fixed journal bugs
fixed the bugs related to the journals, created new script called
journals to keep track of the number of journals collected seperately
from the journal script, which is now on each journal object
independently and no longer operates using a for loop through the
gameobject list of journals, because most of the errors stemmed from the
for loop. Adding journals should not be very complicated, and keeping
track of which journal is picked up should also be an easy addition.